==> notebook1.ipynb <==
import pandas as pd
import os

# Define file paths
scripts_path = "/Users/Carys/Documents/github/show-synth/data/raw/got-scripts.csv"
ratings_path = "/Users/Carys/Documents/github/show-synth/data/raw/Game_of_Thrones.csv"



# load CSVs

scripts_df = pd.read_csv(scripts_path)
ratings_df = pd.read_csv(ratings_path)

print("Scripts DataFrame:")
display(scripts_df.head())

print("\nRatings DataFrame:")
display(ratings_df.head())


# find missing data

print("Scripts dataset info:")
scripts_df.info()
print("\nMissing values (scripts):")
print(scripts_df.isnull().sum())

print("\nRatings dataset info:")
ratings_df.info()
print("\nMissing values (ratings):")
print(ratings_df.isnull().sum())


# remove duplicates

scripts_df.drop_duplicates(inplace=True)
ratings_df.drop_duplicates(inplace=True)

# save clean csvs
scripts_df.to_csv("/Users/Carys/Documents/github/show-synth/data/processed/clean_scripts_initial.csv", index=False)
ratings_df.to_csv("/Users/Carys/Documents/github/show-synth/data/processed/clean_ratings_initial.csv", index=False)

print("First cleaned CSVs saved to data/processed/")


scripts_df['Speaker'] = scripts_df['Speaker'].fillna("Narration")

scripts_df['Speaker'].value_counts().head()
