==> notebook2_preprocessing.ipynb <==
import pandas as pd
import re

scripts_df = pd.read_csv("/Users/Carys/Documents/github/show-synth/data/processed/clean_scripts_initial.csv")

# add row id column 
scripts_df['row_id'] = scripts_df.index


# extract bracketed text from dialogue column 
def extract_brackets(text):
    if pd.isna(text):
        return []
    # find all narration chunks
    return re.findall(r'\[(.*?)\]', text)

scripts_df['narration_text'] = scripts_df['Text'].apply(extract_brackets)


# one row per bracketed text
brackets_df = scripts_df[['row_id', 'Episode', 'Season', 'Show', 'narration_text']].explode('narration_text')

# delete/drop rows where there were no brackets
brackets_df = brackets_df.dropna(subset=['narration_text'])


# save bracketed text as its own csv 

brackets_df.to_csv("/Users/Carys/Documents/github/show-synth/data/processed/bracket_narrations.csv", index=False)
print("Saved bracketed directions to data/processed/bracket_narrations.csv")


# remove all bracketed text to save dialogue only  

import re

def remove_brackets(text):
    if pd.isna(text):
        return text
    # remove the entire [ ... ]
    return re.sub(r'\[[^\]]*\]', '', text)


scripts_df['dialogue_text'] = scripts_df['Text'].apply(remove_brackets)


# separate dataframe for dialogue only
dialogue_only = scripts_df[['dialogue_text']].copy()

dialogue_only = dialogue_only[dialogue_only['dialogue_text'].str.strip() != ""]

# save ONLY the dialogue
dialogue_only.to_csv(
    "/Users/Carys/Documents/github/show-synth/data/processed/clean_scripts_dialogue.csv",
    index=False
)
print("Saved dialogue-only scripts to data/processed/clean_scripts_dialogue.csv")
