==> notebook3_emotionprocessing.ipynb <==
import pandas as pd
from nrclex import NRCLex

# load initial scripts + dialogue
scripts_initial = pd.read_csv("/Users/Carys/Documents/github/show-synth/data/processed/clean_scripts_initial.csv")
dialogue_only = pd.read_csv("/Users/Carys/Documents/github/show-synth/data/processed/clean_scripts_dialogue.csv")

# merge episodes and seasons 

dialogue_only["Episode"] = scripts_initial["Episode"]
dialogue_only["Season"] = scripts_initial["Season"]

# merge diaglogue into one string per episode 

episode_text = (
    dialogue_only
    .groupby(["Season", "Episode"])["dialogue_text"]
    .apply(lambda x: " ".join(x.dropna()))
    .reset_index()
)


# run model and assign frequencies to emotions
def get_emotion_freq(text):
    emo = NRCLex(str(text))
    return emo.affect_frequencies   # proportions for anger, joy, fear, etc.

emotion_dicts = episode_text["dialogue_text"].apply(get_emotion_freq)
emotion_df = pd.DataFrame(list(emotion_dicts)).fillna(0.0)


# attach score to the season and episode

episode_emotions = pd.concat(
    [episode_text[["Season", "Episode"]], emotion_df],
    axis=1
)

# save csv 
episode_emotions.to_csv("/Users/Carys/Documents/github/show-synth/data/processed/episode_emotions.csv", index=False)
print("Saved episode_emotions.csv")

#fix extra column
df = df.drop(columns=["anticip"], errors="ignore")
df.to_csv("data/processed/episode_emotions.csv", index=False)
