==> notebook4_emotionstandard.ipynb <==
import pandas as pd

#load first episode emotions file and show all columns
df = pd.read_csv("data/processed/episode_emotions.csv")

print("Columns before cleaning:\n", df.columns.tolist())


#drop empty columns and reorder them to match the NRCLex order
columns_to_drop = ["anticip"]  
df = df.drop(columns=[col for col in columns_to_drop if col in df.columns])

emotion_cols = [
    "fear", "anger", "anticipation", "trust", "surprise",
    "positive", "negative", "sadness", "disgust", "joy"
]

final_cols = ["Season", "Episode"] + emotion_cols
df = df[final_cols]


#save clean csv
df.to_csv("data/processed/episode_emotions_clean.csv", index=False)

print("\nCleaned and standardized emotion CSV saved!")
print("Final columns:\n", df.columns.tolist())
