==> notebook5_keywordextract.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer

#get season/ep data from iniital scripts and combine with dialogue 
scripts = pd.read_csv("data/processed/clean_scripts_initial.csv")
dialogue = pd.read_csv("data/processed/clean_scripts_dialogue.csv")

dialogue["Season"] = scripts["Season"]
dialogue["Episode"] = scripts["Episode"]


episodes = (
    dialogue
    .groupby(["Season", "Episode"])["dialogue_text"]
    .apply(lambda x: " ".join(x.dropna()))
    .reset_index()
)


vectorizer = TfidfVectorizer(stop_words="english", max_features=2000)
tfidf_matrix = vectorizer.fit_transform(episodes["dialogue_text"])
feature_names = vectorizer.get_feature_names_out()

#function for top keywords 
def get_top_keywords(row, n=10):
    row_data = row.toarray().flatten()
    top_idx = row_data.argsort()[-n:][::-1]
    return [feature_names[i] for i in top_idx]

# top keywords per episode calculation
episodes["top_keywords"] = [
    get_top_keywords(tfidf_matrix[i], n=10)
    for i in range(tfidf_matrix.shape[0])
]


episodes = episodes[["Season", "Episode", "top_keywords"]]

episodes.to_csv("data/processed/episode_topkeywords.csv", index=False)

episodes.head()
