==> notebook6_merge.ipynb <==
#merge csvs

import pandas as pd

emotions = pd.read_csv('data/processed/episode_emotions_clean.csv')

ratings = pd.read_csv('data/processed/clean_ratings_initial.csv')

print(emotions.columns)
print(ratings.columns)


# standardize season numbers

ratings['Season'] = ratings['Season'].apply(lambda x: f"season-{int(x):02d}")
ratings['Episode'] = ratings['No. of Episode (Season)'].apply(lambda x: f"e{int(x):02d}")

emotions['Episode'] = emotions['Episode'].str.extract(r'(e\d+)', expand=False)
emotions['Episode'] = emotions['Episode'].str.replace('e(\d)', r'e0\1', regex=True)



merged = pd.merge(
    emotions,
    ratings,
    on=['Season', 'Episode'],
    how='inner'
)

merged.to_csv("data/processed/episode_full_dataset.csv", index=False)
